# openapi/v1.yaml
openapi: 3.0.3
info:
  title: Project Algebra API
  version: "1.0"
  description: >
    Agentic-commerce control plane REST API. Shares the exact same
    application services as the MCP server (see docs/MCP.md) — this spec
    documents the human/account-management surface plus the same commerce
    endpoints an agent could also reach through MCP.
servers:
  - url: /api/v1
security:
  - agentToken: []
components:
  securitySchemes:
    agentToken:
      type: http
      scheme: bearer
      description: >
        Bearer token minted by POST /agents. Resolved server-side to an
        AgentIdentity; see internal/domain/agent.
    devUserID:
      type: apiKey
      in: header
      name: X-User-ID
      description: >
        DEV-MODE PLACEHOLDER for human-only endpoints (approvals, payment
        sources, privacy profiles). NOT a security boundary — see
        docs/LOCAL_DEVELOPMENT.md. Production requires a real OIDC session.
    integratorToken:
      type: http
      scheme: bearer
      description: >
        Bearer token minted by POST /integrators. Resolved server-side to
        an Integrator (internal/domain/integrator) — a third-party
        application calling the standalone policy-evaluation surface, not
        an AI shopping agent. Never interchangeable with agentToken.
  schemas:
...
    Intent:
      type: object
      properties:
        intent_id: { type: string }
        status:
          type: string
          enum: [DRAFT, DISCOVERING, QUOTED, POLICY_CHECK, POLICY_REJECTED, APPROVAL_REQUIRED, APPROVED, REAPPROVAL_REQUIRED, EXECUTING, AUTHENTICATION_REQUIRED, SUCCEEDED, FAILED, CANCELLED, EXPIRED, PARTIALLY_COMPLETED, MERCHANT_INTERVENTION_REQUIRED, USER_INTERVENTION_REQUIRED]
        selected_quote_id: { type: string }
...
    PolicyDecision:
      type: object
      properties:
        decision: { type: string, enum: [ALLOW, DENY, REQUIRE_APPROVAL] }
        reason_codes: { type: array, items: { type: string } }
        policy_version: { type: string }
    Conditions:
      type: object
      description: >
        A third-party integrator's own budget/category/merchant policy —
        the same shape as internal/domain/policy.Rules, minus the
        crypto-rail fields (no counterpart in this card/wallet-shaped
        request). Required on POST /policy/evaluate-transaction — pass {}
        to explicitly allow everything; Algebra does not assume a default
        budget policy on an integrator's behalf.
      properties:
        max_per_transaction_minor_units: { type: integer, format: int64 }
        max_per_day_minor_units: { type: integer, format: int64 }
        approval_threshold_minor_units: { type: integer, format: int64 }
        allowed_merchants: { type: array, items: { type: string } }
        blocked_merchants: { type: array, items: { type: string } }
        blocked_categories: { type: array, items: { type: string } }
        allowed_payment_profiles: { type: array, items: { type: string } }
        international_requires_approval: { type: boolean }
...
paths:
  /users:
    post:
      summary: Create a user (dev-mode bootstrap — see docs/LOCAL_DEVELOPMENT.md)
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { email: { type: string } }
      responses:
        '201': { description: Created }
  /agents:
    post:
      summary: Mint a new AgentIdentity bearer token (shown exactly once)
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id: { type: string }
                client_id: { type: string }
                name: { type: string }
                permissions:
                  type: array
                  items: { type: string, enum: [shopping.read, shopping.create_intent, shopping.execute, payments.request, orders.read, profiles.read, policy.read] }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties: { agent_id: { type: string }, token: { type: string } }
  /agents/{id}/revoke:
    post:
      summary: Revoke an agent's token immediately — a user action, only the agent's own owner may revoke it
      security: [{ devUserID: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK }, '403': { description: The agent does not belong to the caller } }
  /integrators:
    post:
      summary: >
        Mint a new Integrator bearer token (shown exactly once) — the B2B
        counterpart to POST /agents, for a third-party application calling
        POST /policy/evaluate-transaction. See docs/INTEGRATING.md.
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { name: { type: string, example: Kite } }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties: { integrator_id: { type: string }, token: { type: string } }
  /integrators/{id}/revoke:
    post:
      summary: Revoke an integrator's token immediately
      security: []
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK } }
  /policy/evaluate-transaction:
    post:
      summary: >
        Standalone policy decision for a third-party integrator's own
        transaction — no Algebra purchase intent involved. See
        docs/INTEGRATING.md for the full request/response reference and
        worked examples.
      security: [{ integratorToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [where, how_much, conditions]
              properties:
                who:
                  type: object
                  properties: { user_ref: { type: string, description: "the integrator's own opaque user identifier" } }
                what:
                  type: object
                  properties: { category: { type: string } }
                where:
                  type: object
                  properties:
                    merchant: { type: string }
                    international: { type: boolean }
                how_much:
                  type: object
                  properties:
                    amount_minor_units: { type: integer, format: int64 }
                    currency: { type: string, example: USD }
                    spend_today_minor_units: { type: integer, format: int64, description: "caller-supplied — Algebra keeps no ledger for transactions it never processes" }
                with_what:
                  type: object
                  properties: { payment_ref: { type: string, description: "the integrator's own opaque payment-method reference" } }
                conditions: { $ref: '#/components/schemas/Conditions' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PolicyDecision' }
        '400': { description: Missing conditions, or the integrator token is invalid/revoked }
  /intents:
    post:
      summary: commerce.create_purchase_intent
      parameters:
        - { name: Idempotency-Key, in: header, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    type: object
                    properties: { query: { type: string }, quantity: { type: integer } }
                constraints:
                  type: object
                  properties:
                    max_total_minor_units: { type: integer }
                    currency: { type: string }
                    delivery_profile: { type: string }
                    payment_profile: { type: string }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Intent' } } } }
  /intents/{id}:
    get:
      summary: commerce.get_purchase_intent
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Intent' } } } } }
  /intents/{id}/cancel:
    post:
      summary: commerce.cancel_purchase
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK } }
  /intents/{id}/discover:
    post:
      summary: Run merchant discovery for this intent
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { type: object, properties: { quotes: { type: array, items: { $ref: '#/components/schemas/Quote' } } } } } }
  /intents/{id}/quotes:
    get:
      summary: commerce.get_quotes (read cached quotes without re-running discovery)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { type: object, properties: { quotes: { type: array, items: { $ref: '#/components/schemas/Quote' } } } } } }
  /intents/{id}/select-quote:
    post:
      summary: commerce.select_quote
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        content: { application/json: { schema: { type: object, properties: { quote_id: { type: string } } } } }
      responses: { '200': { description: OK } }
  /intents/{id}/request-purchase:
    post:
      summary: commerce.request_purchase — evaluate policy, may create an Approval
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/PolicyDecision' } } } } }
  /intents/{id}/policy-preview:
    get:
      summary: policy.evaluate_intent — dry run, no side effects
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/PolicyDecision' } } } } }
  /intents/{id}/policy-explain:
    get:
      summary: policy.explain_decision
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/PolicyDecision' } } } } }
  /intents/{id}/execute:
    post:
      summary: Execute an APPROVED intent (refreshes quote, re-checks policy, calls the merchant connector)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: Idempotency-Key, in: header, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  intent_status: { type: string }
                  order: { $ref: '#/components/schemas/Order' }
                  reason: { type: string }
  /intents/{id}/order:
    get:
      summary: commerce.get_order_status
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Order' } } } } }
  /intents/{id}/receipt:
    get:
      summary: commerce.get_receipt
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Order' } } } } }
  /intents/{id}/complete-authentication:
    post:
      summary: Resume an intent parked in AUTHENTICATION_REQUIRED after the user completed 3DS/OTP/UPI on their own device
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                merchant_order_id: { type: string }
                success: { type: boolean }
      responses: { '200': { description: OK } }
  /intents/{id}/cancel-order:
    post:
      summary: Ask the MERCHANT to cancel the placed order (a merchant refusal surfaces as an error; Algebra never cancels locally on its own say-so)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Order' } } } }
        '409': { description: Merchant refused (already shipped, past cancellation window) }
  /intents/{id}/audit:
    get:
      summary: The append-only audit trail for this intent — which agent requested it, what policy decided, what the user approved, what was charged
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK } }
  /intents/{id}/approval:
    get:
      summary: Get the current approval for an intent (what the Approval UI renders before Approve/Reject)
      security: [{ devUserID: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK } }
  /approvals/{id}/approve:
    post:
      summary: The actual "Approve" button behind the Approval UI — a user action (X-User-ID dev placeholder)
      security: [{ devUserID: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK } }
  /approvals/{id}/reject:
    post:
      summary: The "Reject" button
      security: [{ devUserID: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK } }
  /approvals/{id}/reapprove:
    post:
      summary: Re-confirm an approval after REAPPROVAL_REQUIRED (merchant price drifted beyond tolerance)
      security: [{ devUserID: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK } }
  /payment-sources:
    get:
      summary: payments.list_sources
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/PaymentSource' } } } }
    post:
      summary: Add a card via a vault provider_nonce (never a raw PAN) — a user action
      security: [{ devUserID: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                provider_nonce: { type: string }
                alias: { type: string }
                nickname: { type: string }
      responses: { '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/PaymentSource' } } } } }
  /payment-sources/{id}/revoke:
    post:
      summary: Revoke a payment source — only the source's own owner may revoke it
      security: [{ devUserID: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK }, '403': { description: The payment source does not belong to the caller } }
...
  /webhooks/{provider}:
    post:
      summary: Generic inbound webhook receiver — HMAC-verified, deduplicated, rejected if no secret is configured for {provider}
      security: []
      parameters:
        - { name: provider, in: path, required: true, schema: { type: string } }
        - { name: X-Algebra-Signature, in: header, schema: { type: string }, description: "sha256=<hex hmac> — see internal/app/webhook_service.go" }
        - { name: X-Algebra-Event-ID, in: header, schema: { type: string }, description: "optional; falls back to a payload fingerprint for dedup if omitted" }
      responses:
        '200': { description: Accepted (ok:false if this was a duplicate delivery) }
        '401': { description: Signature invalid or no secret configured for this provider }
  /profiles/billing:
    post:
      summary: Create a billing profile (encrypted at rest) — a user action, never exposed via MCP
      security: [{ devUserID: [] }]
      responses: { '201': { description: Created } }